feat(#2827): measure partition extent as the successor gap (owner ruling, option b)
R1: size_source gains a distinct third value successor_gap, so a reader can always
tell a MEASURED extent from an index-supplied one and both from unknown;
unavailable still fails closed at zero bytes. Mixed provenance reports the WEAKEST
component.

SSTableReader::measure_partition_extent + uncompressed_data_section_len bound
[data_offset, successor_offset) — the same authoritative index-layout bound the
single-partition seek uses — with the data-section-length bound for the last
partition. Reachable at the LOGICAL boundary: both weight resolvers already ask
reader-level state there, so the probe does NOT move.

The decision procedure now prices a gap-measured window instead of refusing it;
refusal condition 1 tests INCOMPLETENESS, not provenance. Corrects the false
'BIG resolves both fields' PartitionLoc doc (its tests split to a sibling file to
stay inside the campsite ratchet).
